[lab01] Change double to float

+++ 01-inradius-of-a-right-triangle/main.cpp
@@ -3,17 +3,17 @@
 
 
 int main() {
-    double leg_a, leg_b,  // Катети (a та b)
-           hypotenuse,    // Гіпетенуза (c)
-           radius;        // Радіус вписаного кола
+    float leg_a, leg_b,  // Катети (a та b)
+          hypotenuse,    // Гіпетенуза (c)
+          radius;        // Радіус вписаного кола
 
-    printf("Введіть катет (a): "); scanf("%lf", &leg_a);
-    printf("Введіть гіпотенузу (c): "); scanf("%lf", &hypotenuse);
+    printf("Введіть катет (a): "); scanf("%f", &leg_a);
+    printf("Введіть гіпотенузу (c): "); scanf("%f", &hypotenuse);
 
     leg_b = sqrt(pow(hypotenuse, 2) - pow(leg_a, 2));  // Катет b за теоремою піфагора
     radius = (leg_a + leg_b - hypotenuse) / 2;         // Радіус вписаного кола
 
-    printf("Радіус вписаного кола (r): %lg\n", radius);
+    printf("Радіус вписаного кола (r): %g\n", radius);
 
     return 0;
 }
